requirements.txt
Flask==3.0.0
supabase==2.3.4
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0
schedule==1.2.0
gunicorn==21.2.0
openpyxl==3.1.2
pandas==2.1.4
